+++ src/scene.rs
@@ -48,8 +48,8 @@ impl Scene {
             self.add(obj);
         }
     }
-    pub fn pdf_value(&self, origin: &Vector3<f32>, direction: &Vector3<f32>) -> f32 {
-        let weight = 1.0 / self.objects.len() as f32;
+    pub fn pdf_value(&self, origin: &Vector3<f64>, direction: &Vector3<f64>) -> f64 {
+        let weight = 1.0 / self.objects.len() as f64;
         let mut sum = 0.0;
 
         for object in self.objects.iter() {
@@ -58,7 +58,7 @@ impl Scene {
 
         sum
     }
-    pub fn random(&self, origin: &Vector3<f32>) -> Vector3<f32> {
+    pub fn random(&self, origin: &Vector3<f64>) -> Vector3<f64> {
         let int_size = self.objects.len() as i32;
         // println!("int_size:{}", int_size);
         self.objects[random_int(0, int_size ) as usize].random(origin)
